+++ packages/implementations/react/src/atomic/Avatar/src/index.ts
@@ -1,0 +1,5 @@
+import { Avatar } from "./Avatar";
+
+export { Avatar } from "./Avatar";
+export type { AvatarProps } from "./avatar.types";
+export default Avatar;