+++ packages/styles/src/packages/atomic/avatar/index.ts
@@ -1,0 +1,6 @@
+import { styles } from "./rarui-avatar.css";
+
+export const avatar = {
+  classnames: { ...styles },
+};
+export type { AvatarVariants } from "./rarui-avatar.types";